refactor heuristic

+++ MCTS/MCTSData.py
@@ -17,7 +17,7 @@ class MCTSData:
         self.distances: np.ndarray
         self.food:list
         self.distributions:list[dict]
-        self.max_distance:float
+        self.max_distance:int
         
 
     def get_food_locations(self) -> None:
